videos/requirements.txt
requests
flask
flask-lambda-python36
mysql-connector-python
mysql-connector-python==8.0.11
